Fixes CurrentArtwork
The CurrentArtwork mistakenly displays only the Artwork of the current Playlist's first song. With the change it will display the the current song's Artwork directly after every songschange.

The official CLI Reference tells us: "Index (first item is 0) of the playlist entry in the player playlist. Unless <start> is "-", the first returned instance of this field is equal to start. If <start> is "-", the first returned instance of this field contains the index of the currently playing song in the player playlist."